requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
apscheduler==3.10.1
asyncio==3.4.3
beautifulsoup4==4.12.2
bing_image_downloader==1.1.2
bleach==6.0.0
bs4==0.0.1
cachetools==5.3.0
countryinfo==0.1.2
currencyconverter==0.17.7
dateparser==1.1.8
emoji==2.2.0
emoji-country-flag==1.3.2
feedparser==6.0.10
future==0.18.3
geopy==2.3.0
gpytranslate==1.5.1
hachoir==3.2.0
html2text==2020.1.16
jikanpy==4.3.2
lxml==4.9.2
markdown2==2.4.8
motor==3.1.2
opencv-python-headless==4.7.0.72
pillow==9.5.0
pretty_errors==1.2.25
psutil==5.9.5
psycopg2-binary==2.9.6
pydictionary==2.0.1
pykeyboard==0.1.5
pynewtonmath==1.0.1
pyrate-limiter==2.10.0
pyrogram==2.0.103
python-telegram-bot==12.8
regex==2023.3.23
requests==2.28.2
secureme==0.0.4
search_engine_parser==0.6.8
speedtest-cli==2.1.3
sqlalchemy==1.4.46
telegraph==2.2.0
telethon==1.28.5
tgcrypto==1.2.5
wikipedia==1.4.0
phlogo==0.6
gtts
heroku3
alphabet_detector
wget
nekos.py
yt_dlp
youtube_search
youtube_search_python
openai
pyshorteners
